P.I. Falhada tentativa de atribuição de casas decimais ao campo Distancia
.Falhada tentativa de atribuição de casas decimais ao campo Distancia
.Nota: Tentei com novas anotaçoes e tipos decimal e double: nao reconhece parte decimal quer separado por ponto ou virgula
Tempo: 1h

+++ Project_Trails4Health/Trails4Health/Migrations/20180114200857_Initial.cs
@@ -105,7 +105,7 @@ namespace Trails4Health.Migrations
                     Desativado = table.Column<bool>(nullable: false),
                     Detalhes = table.Column<string>(maxLength: 700, nullable: false),
                     DificuldadeID = table.Column<int>(nullable: false),
-                    Distancia = table.Column<decimal>(nullable: false),
+                    Distancia = table.Column<double>(nullable: false),
                     Fim = table.Column<string>(maxLength: 50, nullable: false),
                     Foto = table.Column<string>(nullable: false),
                     Inicio = table.Column<string>(maxLength: 50, nullable: false),
